Change print_tree to print AST in dot form
Changed the print_tree method to print out graphviz dot code to
draw the tree.  This is still very crude, but it's much easier to
visualize how the tree is being formed.

To support this change the Ast_types.program type was removed, and
now the root program is represented as a Block statement instead of
a list of statements.  This should be a nicer overall
representation anyway.  The parser handles the type conversion.