Make edge insertion consistent
Some of the edge insertion for the GraphViz DOT generation of the
AST was inconsistent.  Some functions assumed the caller made the
edge and some assumed the callee, resulting in a few double or
missing edges.  Now the caller is always required to create a
node and a link to its parent.