Implement compilation of function declarations
Each function declaration is preceded by a Bra call to skip over
it in case the PC approaches it from something prior falling
through it (e.g. at the start of the program).  This is a quicker
hack than actually reloacting all of them after the Hlt instruction
or something.