Add simple global variables
Added simple global variables.  These are implemented as an array
in the bytecode interpreter.  The compiler assigns each name a
unique entry in the array.  Currently all variables are considered
global.  Variables are not being set at an offset of FP on the stack
because they are dynamically created and not required to be declared.